refactor the prerecorded dashboard and how environment variables are passed to the cli app from TUI

+++ tui/controllers/dashboardController.py
@@ -29,33 +29,50 @@ class DashboardController():
 
         self.view = DashboardView(self.model)
         self.loop = urwid.MainLoop(self.view, palette=palette, unhandled_input=self.handle_input)
-        self.loop.run()
         self.model.running = False
+        self.proc = None
+
+        #Line has to be last
+        self.loop.run()
     
+    def getEnvironmentVars(self):
+        my_env = os.environ.copy()
+        my_env["MODE"] = self.model.mode
+        my_env["ALGORITHM"] = self.model.algorithm
+        my_env["INPUT_FILE"] = self.model.inputFile
+        my_env["TARGET_FILE"] = self.model.targetFile
+        my_env["DEVICE"] = self.model.device
+        my_env["SIZE"] = str(self.model.size)
+        my_env["ROLE"] = self.model.role
+        my_env["WAIT_SIZE"] = str(self.model.waitSize)
+        my_env["STEP_SIZE"] = str(self.model.stepSize)
+        my_env["TUI_CONNECTION"] = str(self.model.tuiConnection)
+        return my_env
+
     def run(self):
-        self.logger.info('Running...')
         self.stdout = self.loop.watch_pipe(self.read_pipe)
         self.stderr = self.loop.watch_pipe(self.read_pipe)
-        cwd = get_project_root()
-        connected = False
-        self.proc = subprocess.Popen(['python3', '-m', 'cli', self.model.mode,self.model.algorithm,self.model.inputFile,
-                                                                self.model.targetFile,self.model.device,str(self.model.size),
-                                                                self.model.role,str(self.model.waitSize),str(self.model.stepSize),
-                                                                str(self.model.tuiConnection)],
-                        cwd=cwd,
+        
+        my_env = self.getEnvironmentVars()
+
+        self.proc = subprocess.Popen(['python3', '-m', 'cli', ],
+                        cwd=get_project_root(),
                         stdout=self.stdout,
-                        stderr=self.stderr
+                        stderr=self.stderr,
+                        env=my_env
                         )
         
-        #loop until connected
-        while not connected:
-            try:
-                self.dataClient = Client(('localhost', 5000), authkey=b'secret password')
-                connected = True
-            except ConnectionRefusedError:
-                pass
+        if(self.model.mode == 'live'):
+            connected = False
+            #loop until connected
+            while not connected:
+                try:
+                    self.dataClient = Client(('localhost', 5000), authkey=b'secret password')
+                    connected = True
+                except ConnectionRefusedError:
+                    pass
 
-        self.logger.debug('Connected to Process!')
+            self.logger.debug('Connected to Process!')
         self.loop.set_alarm_in(0, self.refresh)
     
     def read_pipe(self, read_data):
@@ -63,25 +80,26 @@ class DashboardController():
 
     def refresh(self, _loop, data):
         try:
-            #first update all three buffers
-            tuiBufferName = self.dataClient.recv() #receive 'error'
-            while tuiBufferName != 'end buffers':
-                tuiData = self.dataClient.recv()
-                self.logger.debug(f'Appending {tuiData} to buffer {tuiBufferName}')
+            if(self.model.mode == 'live'):
+                #first update all three buffers
+                tuiBufferName = self.dataClient.recv() #receive 'error'
+                while tuiBufferName != 'end buffers':
+                    tuiData = self.dataClient.recv()
+                    self.logger.debug(f'Appending {tuiData} to buffer {tuiBufferName}')
 
-                if(tuiBufferName == 'error'):
-                    self.model.errorBuffer.append([float(tuiData.flat[0])])
-                if(tuiBufferName == 'output'):
-                    self.model.outputBuffer.append([float(tuiData.flat[0])])
-                if(tuiBufferName == 'reference'):
-                    self.model.referenceBuffer.append([float(tuiData.flat[0])])
-                if(tuiBufferName == 'output-error'):
-                    self.model.errorPercentage = tuiData.flat[0]
+                    if(tuiBufferName == 'error'):
+                        self.model.errorBuffer.append([float(tuiData.flat[0])])
+                    if(tuiBufferName == 'output'):
+                        self.model.outputBuffer.append([float(tuiData.flat[0])])
+                    if(tuiBufferName == 'reference'):
+                        self.model.referenceBuffer.append([float(tuiData.flat[0])])
+                    if(tuiBufferName == 'output-error'):
+                        self.model.errorPercentage = tuiData.flat[0]
 
-                tuiBufferName = self.dataClient.recv()
+                    tuiBufferName = self.dataClient.recv()
             
-            self.model.memory = int(self.dataClient.recv())
-            self.model.cpu = float(self.dataClient.recv())
+                self.model.memory = int(self.dataClient.recv())
+                self.model.cpu = float(self.dataClient.recv())
         except EOFError:
             pass
         except Exception as e:
@@ -97,7 +115,9 @@ class DashboardController():
     # Handle key presses
     def handle_input(self, key):
         if key == 'Q' or key == 'q':
-            self.proc.send_signal(signal.SIGINT) 
+            if(self.proc is not None):
+                self.proc.send_signal(signal.SIGINT) 
+
             raise urwid.ExitMainLoop()
         if key == 'R' or key == 'r':
             self.model.running = True
